chunk corrected, PARSE HEADERS problem?
the problem was in content lenght.
need a logic to deal with it, i set it to 0 because my head is gonna explode.
Maybe i find a problem in request factory? if we uncomment the marked code we cant do GET and DELETE, just leave weth break or change logic ?